+++ lib/View/Bottomsheets/add_project_sheet.dart
@@ -256,30 +256,43 @@ class _AddProjectSheetState extends State {
                     CupertinoButton(
                       padding: EdgeInsets.all(0),
                       onPressed: () async {
-                        try {
-                          print('checking internet');
+                        if (_textEditingController.text.isEmpty == false) {
+                          try {
+                            print('checking internet');
 
-                          final result =
-                              await InternetAddress.lookup("example.com");
-                          print('checked internet');
-                          if (result.isNotEmpty &&
-                              result[0].rawAddress.isNotEmpty) {
-                            _loadDataController.projectList.add(Project(
-                                projectName: _textEditingController.text,
-                                projectColor: _loadDataController
-                                    .selectedColorAddProject.value,
-                                projectAvatar: _loadDataController
-                                    .selectedAvatarAddProject.value));
+                            final result =
+                                await InternetAddress.lookup("example.com");
+                            print('checked internet');
+                            if (result.isNotEmpty &&
+                                result[0].rawAddress.isNotEmpty) {
+                              _loadDataController.projectList.add(Project(
+                                  projectName: _textEditingController.text,
+                                  projectColor: _loadDataController
+                                      .selectedColorAddProject.value,
+                                  projectAvatar: _loadDataController
+                                      .selectedAvatarAddProject.value));
 
-                            _database.deleteProjectList(
-                                _loadDataController.currentUserId.value);
+                              _database.deleteProjectList(
+                                  _loadDataController.currentUserId.value);
 
-                            _loadDataController.projectListUpload(
-                                _loadDataController.currentUserId.value);
+                              _loadDataController.projectListUpload(
+                                  _loadDataController.currentUserId.value);
 
-                            Get.back();
-                          } else {
-                            print('WIFI ON NO INTERNET');
+                              Get.back();
+                            } else {
+                              print('WIFI ON NO INTERNET');
+                              Get.showSnackbar(GetSnackBar(
+                                duration: Duration(seconds: 5),
+                                messageText: Text(
+                                  'No internet connection. Please check your internet connection and try again.',
+                                  style: GoogleFonts.poppins(
+                                      color: white,
+                                      fontSize: height * 0.02,
+                                      fontWeight: FontWeight.w400),
+                                ),
+                              ));
+                            }
+                          } on SocketException catch (e) {
                             Get.showSnackbar(GetSnackBar(
                               duration: Duration(seconds: 5),
                               messageText: Text(
@@ -291,11 +304,11 @@ class _AddProjectSheetState extends State {
                               ),
                             ));
                           }
-                        } on SocketException catch (e) {
+                        } else {
                           Get.showSnackbar(GetSnackBar(
                             duration: Duration(seconds: 5),
                             messageText: Text(
-                              'No internet connection. Please check your internet connection and try again.',
+                              'Please add a name of project',
                               style: GoogleFonts.poppins(
                                   color: white,
                                   fontSize: height * 0.02,
